fixes, signed ismp, fee comments
