+++ pallets/api/src/messaging/benchmarking.rs
@@ -11,6 +11,7 @@ use ::xcm::latest::{Junctions, Location};
 use frame_benchmarking::{account, v2::*};
 use frame_support::{dispatch::RawOrigin, traits::Currency, BoundedVec};
 use sp_runtime::traits::{One, Zero};
+use crate::messaging::test_utils::*;
 
 use super::*;
 use crate::Read as _;
@@ -236,43 +237,3 @@ mod messaging_benchmarks {
 	impl_benchmark_test_suite!(Pallet, crate::mock::new_test_ext(), crate::mock::Test);
 }
 
-pub mod utils {
-	use super::*;
-
-	pub fn ismp_get_request() -> GetRequest {
-		GetRequest {
-			source: StateMachine::Polkadot(2000),
-			dest: StateMachine::Polkadot(2001),
-			nonce: 100u64,
-			from: vec![],
-			keys: vec![vec![]],
-			height: 1,
-			context: vec![],
-			timeout_timestamp: 10000,
-		}
-	}
-
-	pub fn ismp_post_request() -> PostRequest {
-		PostRequest {
-			source: StateMachine::Polkadot(2000),
-			dest: StateMachine::Polkadot(2001),
-			nonce: 100u64,
-			from: vec![],
-			to: vec![],
-			timeout_timestamp: 10000,
-			body: vec![],
-		}
-	}
-
-	pub fn ismp_get_response() -> GetResponse {
-		GetResponse { get: ismp_get_request(), values: vec![] }
-	}
-
-	pub fn ismp_post_response() -> PostResponse {
-		PostResponse {
-			post: ismp_post_request(),
-			response: Default::default(),
-			timeout_timestamp: 0,
-		}
-	}
-}
