+++ pallets/api/src/messaging/tests.rs
@@ -7,10 +7,11 @@ use frame_support::{
 	testing_prelude::bounded_vec,
 	weights::Weight,
 };
-use pallet_nfts::{CollectionSetting, MintWitness, WeightInfo as NftsWeightInfoTrait};
 use sp_core::H256;
 
 use crate::{messaging::*, mock::*, Read};
+
+use crate::messaging::test_utils;
 
 pub fn events() -> Vec<Event<Test>> {
 	let result = System::events()
@@ -1091,6 +1092,11 @@ mod ismp_hooks {
 
 	fn handler() -> ismp::Handler<Test> {
 		ismp::Handler::<Test>::new()
+	}
+
+	mod on_timeout {
+		use super::*;
+
 	}
 
 	mod process_response {
