+++ pallets/api/src/mock.rs
@@ -362,6 +362,9 @@ impl Get<StateMachine> for HostStateMachine {
 	}
 }
 
+
+
+
 pub(crate) fn new_test_ext() -> sp_io::TestExternalities {
 	let mut t = frame_system::GenesisConfig::<Test>::default()
 		.build_storage()
